fix: smc test
